requirements.txt
flask
python-dotenv
PyPDF2
google-generativeai==0.3.2
gunicorn
